review service,search searvice

+++ Services/customer-service/src/main/java/com/rumal/customer_service/dto/CustomerResponse.java
@@ -17,7 +17,7 @@ public record CustomerResponse(
         LocalDate dateOfBirth,
         Gender gender,
         CustomerLoyaltyTier loyaltyTier,
-        int loyaltyPoints,
+        long loyaltyPoints,
         List<String> socialProviders,
         boolean active,
         Instant deactivatedAt,
