+++ Services/customer-service/src/main/java/com/rumal/customer_service/dto/analytics/CustomerProfileSummary.java
@@ -8,7 +8,7 @@ public record CustomerProfileSummary(
     String name,
     String email,
     String loyaltyTier,
-    int loyaltyPoints,
+    long loyaltyPoints,
     Instant memberSince,
     boolean active
 ) {}
